Initial setup for breakpoint store
