+++ crates/project/src/debugger.rs
@@ -1,3 +1,4 @@
+pub mod breakpoint_store;
 pub mod dap_command;
 pub mod dap_session;
 pub mod dap_store;
